Revert "double ----int"
This reverts commit a8e6969323c8155193313205788a59a699368a2a.

+++ calculation1.java
@@ -1,5 +1,5 @@
 public class calculation1 {
-    public int calculation1(int n){
-        return (int)Math.pow(2,n);
+    public double calculation1(int n){
+        return Math.pow(2,n);
     }
 }
